+++ src/WPF/ITTV.WPF.MVVM/App.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Windows;
+using ITTV.WPF.MVVM.ViewModels;
 using ITTV.WPF.MVVM.Views;
 using Microsoft.Extensions.DependencyInjection;
