+++ src/WPF/ITTV.WPF.MVVM/Views/MainWindow.xaml
@@ -4,11 +4,18 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:views="clr-namespace:ITTV.WPF.MVVM.Views"
+        xmlns:viewModels="clr-namespace:ITTV.WPF.MVVM.ViewModels"
         mc:Ignorable="d"
         Title="MainWindow" Height="350" Width="525"
-        d:DataContext="{d:DesignInstance views:MainWindowViewModel,
+        d:DataContext="{d:DesignInstance viewModels:MainWindowViewModel,
                                     IsDesignTimeCreatable=True}">
     <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*" />
+            <RowDefinition Height="90" />
+        </Grid.RowDefinitions>
+
+        
         <Grid Grid.Row="1">
                 <Grid.ColumnDefinitions>
                     <ColumnDefinition Width="Auto" />
