+++ src/WPF/ITTV.WPF/ITTV.WPF.csproj
@@ -306,6 +306,9 @@
       <Name>ITTV.WPF.Sources</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="ViewModels" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>
